# perf: reduce heap allocation and improve overall performence