--- a/examples/void-worker/main.go
+++ b/examples/void-worker/main.go
@@ -8,17 +8,16 @@ import (
 )
 
 func main() {
-	start := time.Now()
-	defer func() {
-		fmt.Println("Time taken:", time.Since(start))
-	}()
-
 	w := varmq.NewVoidWorker(func(data int) {
 		fmt.Printf("Processing: %d\n", data)
 		time.Sleep(1 * time.Second)
 	}, 100)
 
 	q := w.BindQueue()
+	start := time.Now()
+	defer func() {
+		fmt.Println("Time taken:", time.Since(start))
+	}()
 	defer q.WaitAndClose()
 	defer fmt.Println("Added jobs")