--- a/result.go
+++ b/result.go
@@ -19,6 +19,8 @@ type EnqueuedJob[R any] interface {
 type EnqueuedGroupJob[T any] interface {
 	// Len returns the number of jobs in the group.
 	Len() int
+	// Wait blocks until all jobs in the group are completed.
+	Wait()
 	// Drain discards the job's result and error values asynchronously.
 	Drain() error
 	// Results returns a channel that will receive the results of the group