--- a/result_channel.go
+++ b/result_channel.go
@@ -10,7 +10,7 @@ import (
 // for concurrent job processing.
 type resultChannel[R any] struct {
 	ch       chan Result[R]
-	consumed atomic.Bool
+	consumed atomic.Value
 }
 
 // newResultChannel creates a new resultChannel with the specified buffer size.
@@ -21,11 +21,12 @@ func newResultChannel[R any](cap int) resultChannel[R] {
 }
 
 func (rc *resultChannel[R]) Read() (<-chan Result[R], error) {
-	if rc.consumed.CompareAndSwap(false, true) {
-		return rc.ch, nil
+	if rc.consumed.Load().(bool) {
+		return nil, errors.New("result channel has already been consumed")
 	}
 
-	return nil, errors.New("result channel has already been consumed")
+	rc.consumed.Store(true)
+	return rc.ch, nil
 }
 
 func (c *resultChannel[R]) Send(result Result[R]) {