- **Refactor**
  - Simplified internal handling of job results and result channels for improved efficiency.
  - Updated benchmark worker creation to use default concurrency settings.
  - Enhanced group job synchronization with unified control for better job counting and completion tracking.
  - Improved timing measurement placement in example usage.
  - Adjusted configuration field types for consistency.
- **Documentation**
  - Clarified example usage by explicitly handling errors in queue operations.

### Old bench
```
BenchmarkQueue_Operations/Add-4               	  806361	      1384 ns/op	     384 B/op	       8 allocs/op
BenchmarkQueue_Operations/AddAll-4            	      1194   1003741 ns/op	  243343 B/op	    7008 allocs/op
```

### New bench
```
BenchmarkQueue_Operations/Add-4               	  928090	      1258 ns/op	     320 B/op	       5 allocs/op
BenchmarkQueue_Operations/AddAll-4            	    1416	    833844 ns/op	  235428 B/op	    5005 allocs/op
```